+++ apps/silverback-gatsby/cypress/integration/constants.ts
@@ -1,4 +1,0 @@
-export const drupalUrl = 'http://localhost:8888';
-export const drupalNodeOpUrl = 'http://localhost:8888/node-op';
-export const previewUrl = 'http://localhost:8000';
-export const siteUrl = 'http://localhost:9000';