+++ packages/npm/@amazeelabs/silverback-playwright/src/gatsby-build/global-setup.ts
@@ -2,25 +2,25 @@ import { $, cd } from 'zx';
 
 import { gatsby } from '../constants';
 import drupalGlobalSetup from '../drupal-only/global-setup';
-import { port } from '../utils';
+import { log, port } from '../utils';
 
 $.verbose = !!process.env.SP_VERBOSE;
 
 export default async function globalSetup() {
+  log('gatsby-build globalSetup start');
+
   await drupalGlobalSetup();
 
   cd(gatsby.path);
-
-  // Gatsby loads env vars from .env file. Create it.
-  await $`source .envrc`;
-
-  await port.killIfUsed(gatsby.fastBuilds.port);
-  await port.killIfUsed(gatsby.port);
+  await port.killIfUsed(gatsby.allPorts);
   await $`yarn clean`;
-  $`yarn fast-builds:serve:local`;
+  // Load env vars right before starting Gatsby so that it sees them.
+  $`source .envrc && yarn fast-builds:serve:local`;
   await port.waitUntilUsed(
     gatsby.port,
     gatsby.timings.retryInterval,
     gatsby.timings.startTimeout,
   );
+
+  log('gatsby-build globalSetup end');
 }