+++ packages/composer/drupal/gatsby_build_monitor/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.5...@-drupal/gatsby_build_monitor@1.3.6) (2021-10-04)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.3.5](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.4...@-drupal/gatsby_build_monitor@1.3.5) (2021-09-28)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.3.4](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.3...@-drupal/gatsby_build_monitor@1.3.4) (2021-09-28)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.3.3](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.2...@-drupal/gatsby_build_monitor@1.3.3) (2021-09-20)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
+## [1.3.2](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.1...@-drupal/gatsby_build_monitor@1.3.2) (2021-09-15)
+
+**Note:** Version bump only for package @-drupal/gatsby_build_monitor
+
+
+
+
+
 ## [1.3.1](https://github.com/AmazeeLabs/silverback-mono/compare/@-drupal/gatsby_build_monitor@1.3.0...@-drupal/gatsby_build_monitor@1.3.1) (2021-07-12)
 
 **Note:** Version bump only for package @-drupal/gatsby_build_monitor