+++ packages/npm/@amazeelabs/react-framework-bridge/CHANGELOG.md
@@ -3,6 +3,138 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.5.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.4.1...@amazeelabs/react-framework-bridge@1.5.0) (2021-10-05)
+
+
+### Features
+
+* added Image.src and Link.href ([dac3cd3](https://github.com/AmazeeLabs/silverback-mono/commit/dac3cd37939b9311f1906adb9c895187582c8d5f))
+
+
+
+
+
+## [1.4.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.4.0...@amazeelabs/react-framework-bridge@1.4.1) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+# [1.4.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.7...@amazeelabs/react-framework-bridge@1.4.0) (2021-10-04)
+
+
+### Features
+
+* a way to get initial html string from Html component ([17bd9dd](https://github.com/AmazeeLabs/silverback-mono/commit/17bd9dd119bb4f21313426152c62066b9f17fa7f))
+
+
+
+
+
+## [1.3.7](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.6...@amazeelabs/react-framework-bridge@1.3.7) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+## [1.3.6](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.5...@amazeelabs/react-framework-bridge@1.3.6) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+## [1.3.5](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.4...@amazeelabs/react-framework-bridge@1.3.5) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+## [1.3.4](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.3...@amazeelabs/react-framework-bridge@1.3.4) (2021-09-15)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+## [1.3.3](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.2...@amazeelabs/react-framework-bridge@1.3.3) (2021-09-14)
+
+
+### Bug Fixes
+
+* **gatsby:** framework bridge link query typing ([afe7943](https://github.com/AmazeeLabs/silverback-mono/commit/afe7943978a5eac82d2d97d43ee54e46bfc8e3f7))
+
+
+
+
+
+## [1.3.2](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.1...@amazeelabs/react-framework-bridge@1.3.2) (2021-09-14)
+
+
+### Bug Fixes
+
+* **gatsby:** default type argument for storybook link builder ([b030ef2](https://github.com/AmazeeLabs/silverback-mono/commit/b030ef2ef476dfc32e2afea726f3a4353dcfe282))
+
+
+
+
+
+## [1.3.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.3.0...@amazeelabs/react-framework-bridge@1.3.1) (2021-09-14)
+
+
+### Bug Fixes
+
+* **gatsby:** added missing buildForm exports ([f258879](https://github.com/AmazeeLabs/silverback-mono/commit/f258879439e5cd43c9b07b5ab9624b298d64b26e))
+
+
+
+
+
+# [1.3.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.2.3...@amazeelabs/react-framework-bridge@1.3.0) (2021-09-14)
+
+
+### Bug Fixes
+
+* **gatsby:** storybook builder unit tests ([9e0fdee](https://github.com/AmazeeLabs/silverback-mono/commit/9e0fdeeafcae8da2474d943926b1298a6eb04c16))
+
+
+### Features
+
+* **gatsby:** add optional query type hints to framework bridge links ([0d7ad9c](https://github.com/AmazeeLabs/silverback-mono/commit/0d7ad9cb83f0c92e90012499ce60284cc6b7c420))
+* **gatsby:** allow to create query-only links (?foo=bar) ([e1505b5](https://github.com/AmazeeLabs/silverback-mono/commit/e1505b52fa379eb621666330976e5a8e5b41a22b))
+* **gatsby:** integrated formik builders for gatsby and storybook ([ae34f11](https://github.com/AmazeeLabs/silverback-mono/commit/ae34f11e1729c0e411bc3f711675c5423e1d06bd))
+
+
+
+
+
+## [1.2.3](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.2.2...@amazeelabs/react-framework-bridge@1.2.3) (2021-09-09)
+
+**Note:** Version bump only for package @amazeelabs/react-framework-bridge
+
+
+
+
+
+## [1.2.2](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.2.1...@amazeelabs/react-framework-bridge@1.2.2) (2021-09-08)
+
+
+### Bug Fixes
+
+* hash urls should be considered relative ([d7e0e06](https://github.com/AmazeeLabs/silverback-mono/commit/d7e0e06ee33553208bac422c9b3ca5e8d14b3ade))
+
+
+
+
+
 ## [1.2.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/react-framework-bridge@1.2.0...@amazeelabs/react-framework-bridge@1.2.1) (2021-09-06)
 
 **Note:** Version bump only for package @amazeelabs/react-framework-bridge