+++ packages/npm/@amazeelabs/silverback-playwright/CHANGELOG.md
@@ -3,6 +3,76 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.3](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.2...@amazeelabs/silverback-playwright@1.3.3) (2021-10-04)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.3.2](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.1...@amazeelabs/silverback-playwright@1.3.2) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.3.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.3.0...@amazeelabs/silverback-playwright@1.3.1) (2021-09-28)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+# [1.3.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.2.2...@amazeelabs/silverback-playwright@1.3.0) (2021-09-15)
+
+
+### Bug Fixes
+
+* ensure env vars are loaded ([b926ad8](https://github.com/AmazeeLabs/silverback-mono/commit/b926ad8a1866dc8a72919d49d16b9fdf2e6b2f7b))
+* kill all gatsby ports to avoid hanging processes ([945a65d](https://github.com/AmazeeLabs/silverback-mono/commit/945a65d84a03cca76a2d8c5f9cd0695c835a3aed))
+* reliable way to run `gatsby develop` ([41b92c6](https://github.com/AmazeeLabs/silverback-mono/commit/41b92c61534f4f64c3fe9d1b4999cc469fffd6d4))
+* use --workers 1 ([5e5ea0c](https://github.com/AmazeeLabs/silverback-mono/commit/5e5ea0c4707672714f45a9517324fa5f94531392))
+
+
+### Features
+
+* introduce --trace flag ([1a9db10](https://github.com/AmazeeLabs/silverback-mono/commit/1a9db10d19b2bf10e0052f34801f035543d4efc0))
+
+
+
+
+
+## [1.2.2](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.2.1...@amazeelabs/silverback-playwright@1.2.2) (2021-09-15)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+## [1.2.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.2.0...@amazeelabs/silverback-playwright@1.2.1) (2021-09-09)
+
+**Note:** Version bump only for package @amazeelabs/silverback-playwright
+
+
+
+
+
+# [1.2.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/silverback-playwright@1.1.0...@amazeelabs/silverback-playwright@1.2.0) (2021-09-09)
+
+
+### Features
+
+* record traces on first retry ([c8485a8](https://github.com/AmazeeLabs/silverback-mono/commit/c8485a84b0b5611ce702cf27a3203c2bf4b5a863))
+
+
+
+
+
 # 1.1.0 (2021-09-06)